+++ Sql/Schema/Stored Procedures/Downloads/Download_Update.sql
@@ -2,7 +2,7 @@
 	DROP PROCEDURE [dbo].[Download_Update]
 GO
 CREATE PROCEDURE [dbo].[Download_Update]
-	@qid int = 0,
+	@qid bigint = 0,
 	@status int = 0
 AS
 	UPDATE DownloadQueue SET status=@status WHERE qid=@qid
